+++ examples/errorHandling/app.js
@@ -3,27 +3,20 @@ import R from 'ramda'
 import F from '../../lib'
 const h = F.h
 
+import submodule from './submodule'
 
-let childs = {
-  submodule: require('./submodule').default,
-}
 
 let moduleDef = F.def({
   name: 'Main',
+  modules: {
+    submodule,
+  },
   init: ({key}) => ({
     key,
     isActive: false,
-    ...F.mergeModels(childs),
   }),
-  load: (ctx, i, Action) => {
-    return R.mapObjIndexed((md, name) => F.createContext(md, {action$: i._childAction(name, md.update)}), childs)
-  },
-  inputs: {
-    _childAction: (ctx, Action, name, update, a) => Action._ChildAction(name, update, a),
-  },
   actions: {
     Toggle: [[], R.evolve({isActive: R.not})],
-    _ChildAction: [[String, R.T, Array], (name, update, a, m) => R.evolve({[name]: update(a)}, m)],
   },
   interfaces: {
     view: ({styles, _md}, i, m) => h('div', {key: m.key, class: {[styles.base]: true}}, [